[PSym] refactor: rename packages
