+++ Src/PCompiler/CompilerCore/Backend/Symbolic/SymbolicCodeGenerator.cs
@@ -2650,15 +2650,17 @@ namespace Plang.Compiler.Backend.Symbolic
             context.WriteLine(output, "package psym.model;");
             context.WriteLine(output);
             context.WriteLine(output, "import psym.commandline.*;");
-            context.WriteLine(output, "import psym.valuesummary.*;");
             context.WriteLine(output, "import psym.runtime.*;");
-            context.WriteLine(output, "import psym.runtime.scheduler.*;");
-            context.WriteLine(output, "import psym.runtime.machine.*;");
             context.WriteLine(output, "import psym.runtime.logger.*;");
+            context.WriteLine(output, "import psym.runtime.machine.*;");
             context.WriteLine(output, "import psym.runtime.machine.buffer.*;");
             context.WriteLine(output, "import psym.runtime.machine.eventhandlers.*;");
+            context.WriteLine(output, "import psym.runtime.machine.events.*;");
+            context.WriteLine(output, "import psym.runtime.scheduler.*;");
             context.WriteLine(output, "import psym.runtime.values.*;");
             context.WriteLine(output, "import psym.utils.*;");
+            context.WriteLine(output, "import psym.utils.serialize.*;");
+            context.WriteLine(output, "import psym.valuesummary.*;");
             context.WriteLine(output, "import java.util.List;");
             context.WriteLine(output, "import java.util.ArrayList;");
             context.WriteLine(output, "import java.util.Map;");
