+++ Src/PRuntimes/PSymRuntime/src/main/java/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Manifest-Version: 0.1
-Main-Class: psym.commandline.PSym
+Main-Class: psym.PSym
