+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/EntryPoint.java
@@ -1,12 +1,18 @@
-package psym.commandline;
+package psym;
 
+import psym.commandline.PSymConfiguration;
 import psym.runtime.Concretizer;
 import psym.runtime.logger.*;
 import psym.runtime.scheduler.SearchScheduler;
 import psym.runtime.scheduler.ReplayScheduler;
 import psym.runtime.scheduler.symmetry.SymmetryMode;
 import psym.runtime.scheduler.symmetry.SymmetryTracker;
-import psym.utils.*;
+import psym.utils.exception.BugFoundException;
+import psym.utils.exception.LivenessException;
+import psym.utils.exception.MemoutException;
+import psym.utils.monitor.MemoryMonitor;
+import psym.utils.monitor.TimeMonitor;
+import psym.utils.monitor.TimedCall;
 import psym.valuesummary.Guard;
 import psym.valuesummary.solvers.SolverEngine;
