+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/machine/buffer/EventQueue.java
@@ -1,7 +1,7 @@
 package psym.runtime.machine.buffer;
 
-import psym.runtime.Event;
-import psym.runtime.Message;
+import psym.runtime.machine.events.Event;
+import psym.runtime.machine.events.Message;
 import psym.runtime.logger.TraceLogger;
 import psym.runtime.machine.Machine;
 import psym.runtime.scheduler.Scheduler;
