+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/machine/eventhandlers/DeferEventHandler.java
@@ -1,7 +1,7 @@
 package psym.runtime.machine.eventhandlers;
 
-import psym.runtime.Event;
-import psym.runtime.Message;
+import psym.runtime.machine.events.Event;
+import psym.runtime.machine.events.Message;
 import psym.runtime.machine.Machine;
 import psym.valuesummary.Guard;
 import psym.valuesummary.PrimitiveVS;
