+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/machine/eventhandlers/IgnoreEventHandler.java
@@ -1,6 +1,6 @@
 package psym.runtime.machine.eventhandlers;
 
-import psym.runtime.Event;
+import psym.runtime.machine.events.Event;
 import psym.runtime.machine.Machine;
 import psym.valuesummary.Guard;
 import psym.valuesummary.UnionVS;
