+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/explicit/StateCachingMode.java
@@ -1,4 +1,4 @@
-package psym.utils;
+package psym.runtime.scheduler.explicit;
 
 public enum StateCachingMode {
     None,
