+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/explicit/choiceorchestration/ChoiceOrchestratorRandom.java
@@ -1,6 +1,6 @@
 package psym.runtime.scheduler.explicit.choiceorchestration;
 
-import psym.utils.RandomNumberGenerator;
+import psym.utils.random.RandomNumberGenerator;
 import psym.valuesummary.ValueSummary;
 
 import java.util.Collections;
