+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/utils/exception/BugFoundException.java
@@ -1,4 +1,4 @@
-package psym.utils;
+package psym.utils.exception;
 
 import psym.valuesummary.Guard;
