+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/utils/monitor/TimeMonitor.java
@@ -1,4 +1,4 @@
-package psym.utils;
+package psym.utils.monitor;
 
 import lombok.Getter;
