+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/utils/random/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package psym.utils;
+package psym.utils.random;
 
 import java.util.Random;
